+++ src/app/component/edition-details/edition-details.component.ts
@@ -28,9 +28,11 @@ export class EditionDetailsComponent implements OnInit {
   }
 
     deleteEdition() {
-      this.editionService.deleteEdition(this.id).subscribe(response => {
-      this.router.navigateByUrl('edition');
-      });
+      if (confirm('Are you sure to delete this edition ?')) {
+        this.editionService.deleteEdition(this.id).subscribe(response => {
+        this.router.navigateByUrl('edition');
+        });
+    }
     }
 }
